+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 (unreleased)
+
+- Teams template variable support
+- Filter projects by team in variables
+
 ## 1.0.1 (2022-02-07)
 
 - Docs update to plugin requirements.